requirements.txt
uvicorn[standard]==0.15.0
fastapi==0.70.*
psycopg2-binary
pydantic